requirements.txt
instagrapi==2.0.1
boto3==1.34.34
python-dotenv==1.0.1
Pillow==10.2.0
pytest==8.0.2
requests==2.31.0
python-dateutil==2.8.2 